Add support for resending messages and a new endpoint for initiating the broadcast loop
The commit adds a new endpoint init_room to the Rocket server, which starts the broadcast loop in the Room state managed by the server. The broadcast loop sends all the messages received on the outgoing_stream to all the servers by resending the message with Resended appended to it. The commit also adds a new OutgoingSink struct that abstracts a sink that can send messages to the outgoing_stream. The OutgoingSink struct is used to manage the outgoing message sink and is included in the Rocket state as a managed object.